third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-picture-source-src-and-srcset-expected.txt

CONSOLE WARNING: <source src> with a <picture> parent is invalid and therefore ignored. Please use <source srcset> instead.
